+++ src/WordLookup.h
@@ -11,4 +11,8 @@ bool ShowChineseWordLookupAt(MainWindow* win, TextSelection* ts, EngineBase* eng
 bool ShowEbookWordLookupAt(MainWindow* win, DisplayModel* dm, int pageNo, PointF pagePt, Point screenPos);
 void ShowWordLookup(MainWindow* win, const char* word, Point screenPos);
 void CloseWordLookup();
+bool IsWordLookupVisible();
 void RefreshWordLookupTheme();
+bool IsOfflineDictionaryAvailable();
+bool CanLookupSelectionInTab(WindowTab* tab);
+bool LookupSelectionInTab(MainWindow* win, WindowTab* tab);